Add  object utility function
Very often in Javascript, one comes across a situation where an conditional block of code has to be executed, based on whether the value of a variable is not null, undefined or NaN i.e. check for the variable's existence. Here, I have used the ,  and  to create an abstraction called  which checks for this popular composite condition.